+++ src/rust/engine/remote_provider/remote_provider_opendal/src/byte_store_tests.rs
@@ -10,7 +10,7 @@ use remote_provider_traits::{ByteStoreProvider, RemoteOptions};
 use testutil::data::TestData;
 use testutil::file::mk_tempfile;
 
-use super::base_opendal::Provider;
+use crate::Provider;
 
 const BASE: &str = "opendal-testing-base";
